MOD: Added function to create a set of member id's.
Spamclusters contain the whole path of their members in a set. This is
importent for creating symbolic links, when the option to not copy all
files is chosen. The actual ID of a member is it's file name. So when
the resulting set will contain only the file names without their path.
This helps when comparing clusters.